db4o.net update
git-svn-id: http://source.db4o.com/db4o/trunk/db4o.net@14589 915b37e7-0715-0410-8bc3-b14b59444072

+++ Db4oUnit.Extensions/Db4oUnit.Extensions/Fixtures/AbstractFileBasedDb4oFixture.cs
@@ -37,7 +37,10 @@ namespace Db4oUnit.Extensions.Fixtures
 
 		protected override void DoClean()
 		{
-			_databaseFile.Delete();
+			if (_databaseFile.Exists())
+			{
+				_databaseFile.Delete();
+			}
 		}
 	}
 }
